New folder/src/requirements.txt
flask==1.1.1
Pillow==6.2.1